+++ src/main/java/view/table/TableController.java
@@ -407,9 +407,9 @@ public class TableController {
         return groupController;
     }
 
-    public int getCountPerHG(String hg, String group, int colIndexHG, int colIndexGroup){
+    public double getCountPerHG(String hg, String group, int colIndexHG, int colIndexGroup){
 
-        int count = 0;
+        double count = 0.0;
         ObservableList<ObservableList> selection = getSelectedRows();
         for(int i = 0; i < selection.size(); i++){
             ObservableList list = selection.get(i);
